+++ CHANGELOG.md
@@ -8,6 +8,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.3.0] - Unreleased
 
 ### Added
+- server-side authentication for the status-reviews page ([216](https://github.com/boarlabsxyz/Kavoon/pull/216))
 - robots.txt for search engine crawlers ([200](https://github.com/boarlabsxyz/Kavoon/pull/200))
 - google html file for verification website's ownership ([199](https://github.com/boarlabsxyz/Kavoon/pull/199))
 - SonarQube cloud test coverage ([196](https://github.com/boarlabsxyz/Kavoon/pull/196))